initial refactor of state, wip
