+++ src/polyfem/CMakeLists.txt
@@ -22,3 +22,4 @@ add_subdirectory(quadrature)
 add_subdirectory(solver)
 add_subdirectory(time_integrator)
 add_subdirectory(utils)
+add_subdirectory(output)
