+++ src/polyfem/output/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(SOURCES
+	Evaluator.cpp
+	OutData.cpp
+)
+
+prepend_current_path(SOURCES)
+polyfem_set_source_group(${SOURCES})
+target_sources(polyfem PRIVATE ${SOURCES})
